Harden fzf-native tests & CI
Follow-up to the UTF-8-on-main port (767d030), closing test-suite gaps the
correctness survey found and fixing build regressions the port introduced.

Build/CI:
- Run the UTF-8 elisp suite: `make test` now runs both fzf-native-test.el
  and fzf-native-utf8-test.el (it was orphaned, running nowhere).
- Add a `ctest-asan` target (ASan+UBSan over the C unit tests, with
  UBSAN_OPTIONS=halt_on_error=1 so UBSan findings fail) and a non-Windows
  CI step running it.
- Fix the C-test build broken by the port: ctest targets now add
  -Iutf8proc-2.10.0 and link libutf8proc.a (fzf.c pulls in utf8proc via
  utf8_char_index.h), and fzf-additions-test.c passes the slab to the
  now-3-arg fzf_has_match.

Tests:
- Add exact-value scoring oracles (independently hand-derived from the fzf
  constants, not captured from the binary) for the suffix, equal, and OR
  operators (ASCII) and for equal/prefix/suffix on UTF-8 strings. The OR
  test pins first-match (not max) semantics.
- Add UTF-8 differential cases to fzf-additions-test.c exercising the
  _utf8 deferral path in fzf_has_match (the gap that let the filter-only
  UTF-8 bug through; mutation-tested to fail if the deferral regresses).
- Add an elisp filter-only fast-path guard: asserts the active path returns
  the same set as the full scorer for ASCII/UTF-8/inverted queries.

Fix:
- Silence the -Wsign-compare warning in utf8_char_index.h (guarded cast;
  behavior-neutral). Rebuild the arm64 module.

Verified: elisp 79/79 + 16/16, make ctest clean, make ctest-asan
sanitizer-clean.

+++ utf8_char_index.h
@@ -71,7 +71,9 @@ static inline utf8_char_map_t* utf8_build_char_map(const char *str, size_t byte_
         if (bytes <= 0) break;
         
         // All bytes in this character map to the same char position
-        for (size_t i = 0; i < bytes; i++) {
+        // bytes is guaranteed > 0 here (negative/zero returns broke out above),
+        // so the cast to size_t cannot wrap to a huge value.
+        for (size_t i = 0; i < (size_t)bytes; i++) {
             map->byte_to_char[byte_pos + i] = char_pos;
         }
